+++ jest.config.js
@@ -1,14 +1,10 @@
-module.exports = {
+export default {
   testEnvironment: 'jsdom',
   setupFilesAfterEnv: [
     '<rootDir>/jest.setup.js',
   ],
   transformIgnorePatterns: [
     'node_modules/(?!(@mapbox/mapbox-gl-draw|geodesy-fn)/)'
-  ],
-  reporters: [
-    'default',
-    'jest-simple-summary'
   ],
   collectCoverage: true,
   collectCoverageFrom: [