+++ src/index.js
@@ -1,2 +1,2 @@
 export * from './modes/index';
-export * from './utils/circle_geojson';
+export * from './utils/circle_geojson.js';