+++ src/modes/draw_circle.js
@@ -1,10 +1,10 @@
-import * as CommonSelectors from '@mapbox/mapbox-gl-draw/src/lib/common_selectors';
-import doubleClickZoom from '@mapbox/mapbox-gl-draw/src/lib/double_click_zoom';
-import * as Constants from '../constants';
-import { createCircle, getCircleCenter } from '../utils/circle_geojson';
-import { distance, initialBearing } from '../utils/geodesy';
-import createGeodesicGeojson from '../utils/create_geodesic_geojson';
-import dragPan from '../utils/drag_pan';
+import * as CommonSelectors from '@mapbox/mapbox-gl-draw/src/lib/common_selectors.js';
+import doubleClickZoom from '@mapbox/mapbox-gl-draw/src/lib/double_click_zoom.js';
+import * as Constants from '../constants.js';
+import { createCircle, getCircleCenter } from '../utils/circle_geojson.js';
+import { distance, initialBearing } from '../utils/geodesy.js';
+import createGeodesicGeojson from '../utils/create_geodesic_geojson.js';
+import dragPan from '../utils/drag_pan.js';
 
 const DrawCircleGeodesic = {};