+++ src/modes/static.js
@@ -1,4 +1,4 @@
-import createGeodesicGeojson from '../utils/create_geodesic_geojson';
+import createGeodesicGeojson from '../utils/create_geodesic_geojson.js';
 
 // copied from https://github.com/mapbox/mapbox-gl-draw-static-mode
 const StaticGeodesic = {};